fini score

+++ genetique.h
@@ -25,7 +25,7 @@ typedef struct s_individu_v2 individu_v2;
 
 struct s_population_v2 {
     grille* g;
-    int* score;
+    double* score;
     int taille;
     individu_v2* t;
 };
@@ -42,7 +42,7 @@ void free_population(population* p);
 individu best(population* p, float eco, float env);
 population_v2* creer_population_v2(int n);
 population_v2* next_generation_v2(population_v2* pop);
-float moyenne_v2(population_v2* p);
+double moyenne_v2(population_v2* p);
 void free_population_v2(population_v2* p);
 grille* best_v2(population_v2* p, float eco, float env);
 #endif // GENETIQUE_H
